requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2
pydantic
alembic
python-dotenv
